remove NukkitBiome
